Extract repository from test file
